add: json analyzed

+++ SOPServer.Service/BusinessModels/ItemModels/BulkItemRequestAutoModel.cs
@@ -9,19 +9,19 @@ namespace SOPServer.Service.BusinessModels.ItemModels
 {
     public class BulkItemRequestAutoModel
     {
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public List<string> ImageURLs { get; set; }
     }
 
     public class BulkItemRequestManualModel
     {
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public List<BulkItemModel> ItemsUpload { get; set; }
     }
 
     public class BulkItemModel
     {
         public string ImageURLs { get; set; }
-        public int CategoryId { get; set; }
+        public long CategoryId { get; set; }
     }
 }
